fix: name updates
