+++ composer.json
@@ -38,6 +38,7 @@
 		"name": "Upsnap",
         "handle": "upsnap",
         "hasCpSection": true,
+        "basePath": "src",
 		"changelogUrl": "https://github.com/Appfoster/upsnap/master/CHANGELOG.md",
         "documentationUrl": "https://github.com/Appfoster/upsnap/wiki",
 		"class": "appfoster\\upsnap\\Upsnap"
