+++ src/assetbundles/upsnap/BrokenLinksAsset.php
@@ -1,14 +1,14 @@
 <?php
-namespace appfoster\upsnap\assetbundles\monitor;
+namespace appfoster\upsnap\assetbundles\upsnap;
 
 use craft\web\AssetBundle;
 use craft\web\assets\cp\CpAsset;
 
-class LighthouseAsset extends AssetBundle
+class BrokenLinksAsset extends AssetBundle
 {
     public function init()
     {
-        $this->sourcePath = "@appfoster/upsnap/assetbundles/monitor/dist";
+        $this->sourcePath = "@appfoster/upsnap/assetbundles/upsnap/dist";
 
 
         $this->depends = [
@@ -16,11 +16,11 @@ class LighthouseAsset extends AssetBundle
         ];
 
         $this->css = [
-            'css/lighthouse-scores.css',
+            'css/broken-links.css',
         ];
 
         $this->js = [
-            'js/lighthouseScore.js',
+            'js/brokenLinks.js',
         ];
 
         parent::init();
