+++ src/assetbundles/upsnap/LighthouseAsset.php
@@ -1,14 +1,14 @@
 <?php
-namespace appfoster\upsnap\assetbundles\monitor;
+namespace appfoster\upsnap\assetbundles\upsnap;
 
 use craft\web\AssetBundle;
 use craft\web\assets\cp\CpAsset;
 
-class BrokenLinksAsset extends AssetBundle
+class LighthouseAsset extends AssetBundle
 {
     public function init()
     {
-        $this->sourcePath = "@appfoster/upsnap/assetbundles/monitor/dist";
+        $this->sourcePath = "@appfoster/upsnap/assetbundles/upsnap/dist";
 
 
         $this->depends = [
@@ -16,11 +16,11 @@ class BrokenLinksAsset extends AssetBundle
         ];
 
         $this->css = [
-            'css/broken-links.css',
+            'css/lighthouse-scores.css',
         ];
 
         $this->js = [
-            'js/brokenLinks.js',
+            'js/lighthouseScore.js',
         ];
 
         parent::init();
