+++ src/assetbundles/upsnap/ReachabilityAsset.php
@@ -1,25 +1,25 @@
 <?php
-namespace appfoster\upsnap\assetbundles\monitor;
+namespace appfoster\upsnap\assetbundles\upsnap;
 
 use craft\web\AssetBundle;
 use craft\web\assets\cp\CpAsset;
 
-class MixedContentAsset extends AssetBundle
+class ReachabilityAsset extends AssetBundle
 {
     public function init()
     {
-        $this->sourcePath = "@appfoster/upsnap/assetbundles/monitor/dist";
+        $this->sourcePath = "@appfoster/upsnap/assetbundles/upsnap/dist";
 
         $this->depends = [
             CommonAsset::class,
         ];
 
         $this->css = [
-            'css/mixed-content.css',
+            'css/reachability.css',
         ];
 
         $this->js = [
-            'js/mixedContent.js',
+            'js/reachability.js',
         ];
 
         parent::init();
