+++ src/models/Settings.php
@@ -5,7 +5,7 @@ namespace appfoster\upsnap\models;
 use craft\base\Model;
 
 /**
- * Site Monitor Settings Model
+ * Upsnap Settings Model
  */
 class Settings extends Model
 {
