+++ src/templates/broken-links/_index.twig
@@ -2,7 +2,7 @@
 {% import '_includes/forms' as forms %}
 
 {% set title = "Broken Links"|t('upsnap') %}
-{% do view.registerAssetBundle("appfoster\\upsnap\\assetbundles\\monitor\\BrokenLinksAsset") %}
+{% do view.registerAssetBundle("appfoster\\upsnap\\assetbundles\\upsnap\\BrokenLinksAsset") %}
 
 {% block actionButton %}
 	<div class="buttons">
