+++ src/templates/domain-check/_index.twig
@@ -2,7 +2,7 @@
 {% import '_includes/forms' as forms %}
 
 {% set title = "Domain Check"|t('upsnap') %}
-{% do view.registerAssetBundle("appfoster\\upsnap\\assetbundles\\monitor\\DomainCheckAsset") %}
+{% do view.registerAssetBundle("appfoster\\upsnap\\assetbundles\\upsnap\\DomainCheckAsset") %}
 
 {% block actionButton %}
 	<div class="buttons">
