+++ src/templates/lighthouse/_index.twig
@@ -2,7 +2,7 @@
 {% import '_includes/forms' as forms %}
 
 {% set title = "Lighthouse Scores"|t('upsnap') %}
-{% do view.registerAssetBundle("appfoster\\upsnap\\assetbundles\\monitor\\LighthouseAsset") %}
+{% do view.registerAssetBundle("appfoster\\upsnap\\assetbundles\\upsnap\\LighthouseAsset") %}
 
 
 {% block actionButton %}
