+++ src/templates/mixed-content/_index.twig
@@ -2,7 +2,7 @@
 {% import '_includes/forms' as forms %}
 
 {% set title = "Mixed Content Check"|t('upsnap') %}
-{% do view.registerAssetBundle("appfoster\\upsnap\\assetbundles\\monitor\\MixedContentAsset") %}
+{% do view.registerAssetBundle("appfoster\\upsnap\\assetbundles\\upsnap\\MixedContentAsset") %}
 
 {% block actionButton %}
 	<div class="buttons">
