+++ src/templates/reachability/_index.twig
@@ -1,11 +1,11 @@
 {% extends '_layouts/cp' %}
 {% import '_includes/forms' as forms %}
 
-{% set title = "Site Monitor"|t('upsnap') %}
+{% set title = "Reachability"|t('upsnap') %}
 {% set tabs = {
     status: { label: 'Current Status', url: url('upsnap/reachability') },
 } %}
-{% do view.registerAssetBundle("appfoster\\upsnap\\assetbundles\\monitor\\ReachabilityAsset") %}
+{% do view.registerAssetBundle("appfoster\\upsnap\\assetbundles\\upsnap\\ReachabilityAsset") %}
 
 {% block actionButton %}
 	<div class="buttons">
