+++ src/templates/reachability/history.twig
@@ -2,7 +2,7 @@
 {% import '_includes/forms' as forms %}
 
 {% set title = "Uptime Monitoring"|t('upsnap') %}
-{% do view.registerAssetBundle("appfoster\\upsnap\\assetbundles\\monitor\\ReachabilityAsset") %}
+{% do view.registerAssetBundle("appfoster\\upsnap\\assetbundles\\upsnap\\ReachabilityAsset") %}
 
 {% set tabs = {
     status: { label: 'Current Status', url: url('upsnap/reachability') },
