+++ src/templates/security-certificates/_index.twig
@@ -1,7 +1,7 @@
 {% extends '_layouts/cp' %}
 {% import '_includes/forms' as forms %}
 
-{% do view.registerAssetBundle("appfoster\\upsnap\\assetbundles\\monitor\\SecurityCertificatesAsset") %}
+{% do view.registerAssetBundle("appfoster\\upsnap\\assetbundles\\upsnap\\SecurityCertificatesAsset") %}
 
 
 {% set title = "Security Certificates"|t('upsnap') %}
